fix: 전수조사 확정 결함 3건 — 만남 자동종료 int4 오버플로·연락처 입력 언마운트·순위 가드 무력
1) auto_close_old_meetings()의 `30 * 86400000`이 int4 곱셈이라
   2,592,000,000 > int4 상한(2,147,483,647) → "integer out of range"로
   매일 05:00 KST 크론이 실행될 때마다 실패했다. 30일 만남 자동 종료가
   한 번도 동작한 적이 없고, 사용자가 status=dating에 영구히 갇혀
   배치 매칭 대상(status=active)에서 계속 제외돼 왔다.
   → 피연산자를 bigint로 승격. 002는 신규 설치용, 011은 기존 DB 재정의용.
   ※ 011은 Supabase SQL Editor 수동 적용 전까지 운영 DB에 반영되지 않는다.

2) Matches.tsx의 연락처 입력이 `{!contactInput[m.id] && (<input .../>)}`
   구조라 한 글자를 치는 순간 자기 자신을 언마운트했다. 매칭 수락이라는
   서비스 최대 전환 지점에서 연락처를 1글자만 입력할 수 있었다.
   → 조건부 래퍼 제거.

3) ranking 라우트의 하루 1회 가드가 `.select("1")`이었다. PostgREST는
   이를 컬럼 "1" 조회로 해석해 에러를 내고, error를 구조분해에서 버려
   already=null이 되어 가드가 항상 통과했다. 게다가 rankings PK가
   (user_id, cycle_date, target_id)라 한 사이클에 여러 행이 존재하므로
   maybeSingle()은 다중행 에러도 낸다.
   → count/head 조회로 교체하고 에러를 명시적으로 처리.

검증: npm ci && npm run lint && npm run build 통과(exit 0).
lint 경고 3건은 전부 기존 것이며 수정 파일에는 없다.

+++ supabase/migrations/002_cron.sql
@@ -28,7 +28,7 @@ language sql as $$
     from meetings mt
     join matches m on m.id = mt.match_id
     where mt.status='active'
-      and mt.started_at < (extract(epoch from now()) * 1000)::bigint - 30 * 86400000
+      and mt.started_at < (extract(epoch from now()) * 1000)::bigint - 30::bigint * 86400000
   ),
   upd_meet as (
     update meetings set status='closed',
